requirements-dev.txt
pytest
pytest-asyncio
websocket-client
playwright
pytest-playwright
